# swagger.yaml
openapi: 3.0.0
paths:
...
  /brands:
    get:
      operationId: getBrands
      parameters:
        - name: pageSize
          required: false
          in: query
          description: Number of items per page
          schema:
            example: 20
            type: number
        - name: pageIndex
          required: false
          in: query
          description: Page index (starts from 0)
          schema:
            example: 0
            type: number
        - name: order
          required: false
          in: query
          description: Sort order
          schema:
            example: asc
            type: string
            enum:
              - asc
              - desc
        - name: orderBy
          required: false
          in: query
          description: Field to order by
          schema:
            example: createdAt
            type: string
            enum:
              - createdAt
              - name
              - updatedAt
        - name: keyword
          required: false
          in: query
          description: Search keyword
          schema:
            example: admin
            type: string
      responses:
        "200":
          description: Retrieve a list of brands with pagination.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageDto"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/BrandWithBrandTranslationsResponseDto"
                    required:
                      - data
      summary: Get a list of brands
      tags: &a8
        - Brands
    post:
      operationId: createBrand
      parameters:
        - name: Authorization
          in: header
          description: Bearer auth token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandRequestDto"
      responses:
        "200":
          description: Creates a new brand with the provided details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBrandResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                example:
                  statusCode: 401
                  message: Unauthorized.
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                example:
                  statusCode: 403
                  message: Forbidden.
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Create a new brand
      tags: *a8
  /brands/{id}:
    get:
      operationId: getBrandById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            $ref: "#/components/schemas/BrandIdParamDto"
      responses:
        "200":
          description: Retrieves a specific brand by its ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandWithBrandTranslationsResponseDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                example:
                  statusCode: 400
                  message:
                    - field: field
                      message: message
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                example:
                  statusCode: 404
                  message: Not Found.
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                example:
                  statusCode: 422
                  message: Unprocessable content.
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example:
                  statusCode: 500
                  message: Internal server error.
      summary: Get a brand by ID
      tags: *a8
    put:
      operationId: updateBrand
      parameters:
        - name: id
          required: true
          in: path
          description: Unique identifier for the brand
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrandRequestDto"
      responses:
        "200":
          description: ""
      tags: *a8
    delete:
      operationId: deleteBrand
      parameters:
        - name: id
          required: true
          in: path
          description: Unique identifier for the brand
          schema:
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteBrandRequestDto"
      responses:
        "200":
          description: ""
      tags: *a8
...
info:
  title: Ecom API
  description: The Ecom API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    PaginationResponseDto:
      type: object
      properties:
        totalPages:
          type: number
          description: Total number of pages
          example: 5
        totalItems:
          type: number
          description: Total number of items
          example: 100
        pageSize:
          type: number
          description: Number of items per page
          example: 20
        pageIndex:
          type: number
          description: Current page index (starts from 0)
          example: 0
      required:
        - totalPages
        - totalItems
        - pageSize
        - pageIndex
    PageDto:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/PaginationResponseDto"
      required:
        - pagination
    LanguageResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Language code (ISO 639-1)
          example: en
        name:
          type: string
          description: Language name
          example: English
      required:
        - id
        - name
...
    BrandTranslationWithLanguageResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the brand translation
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
        name:
          type: string
          description: Name of the brand in the specified language
          example: Brand Name
        description:
          type: string
          description: Description of the brand in the specified language
          example: This is a brand description.
        language:
          description: Language details for the translation
          allOf:
            - $ref: "#/components/schemas/LanguageResponseDto"
      required:
        - id
        - name
        - description
        - language
    BrandWithBrandTranslationsResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the brand
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
        logo:
          type: string
          description: URL to the brand's logo image
          example: https://example.com/logos/brand-logo.png
          format: url
        name:
          type: string
          description: Name of the brand
          example: Brand Name
        brandTranslations:
          description: Translations of the brand in different languages
          type: array
          items:
            $ref: "#/components/schemas/BrandTranslationWithLanguageResponseDto"
      required:
        - id
        - logo
        - name
        - brandTranslations
    BrandIdParamDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the brand
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
      required:
        - id
    CreateBrandRequestDto:
      type: object
      properties:
        logo:
          type: string
          description: URL to the brand's logo image
          example: https://example.com/logos/brand-logo.png
        name:
          type: string
          description: Name of the brand
          example: Brand Name
        brandTranslationIds:
          description: brandTranslationIds
          example: &a17
            - 123e4567-e89b-12d3-a456-426614174000
            - 223e4567-e89b-12d3-a456-426614174001
          type: array
          items:
            type: string
      required:
        - logo
        - name
    CreateBrandResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the brand
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
        logo:
          type: string
          description: URL to the brand's logo image
          example: https://example.com/logos/brand-logo.png
          format: url
        name:
          type: string
          description: Name of the brand
          example: Brand Name
        brandTranslations:
          description: Translations of the brand in different languages
          type: array
          items:
            $ref: "#/components/schemas/BrandTranslationWithLanguageResponseDto"
      required:
        - id
        - logo
        - name
        - brandTranslations
    UpdateBrandRequestDto:
      type: object
      properties:
        logo:
          type: string
          description: URL to the brand's logo image
          example: https://example.com/logos/brand-logo.png
        name:
          type: string
          description: Name of the brand
          example: Brand Name
        brandTranslationIds:
          description: brandTranslationIds
          example: *a17
          type: array
          items:
            type: string
    DeleteBrandRequestDto:
      type: object
      properties:
        isHardDelete:
          type: boolean
          description: Whether to hard delete the brand
          example: false
          default: false
...